// app/lib/pocketbase/db-types.ts
export enum Collections {
	Authorigins = "_authOrigins",
	Externalauths = "_externalAuths",
	Mfas = "_mfas",
	Otps = "_otps",
	Superusers = "_superusers",
	Categories = "categories",
	Entries = "entries",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string

// System fields
export type BaseSystemFields<T = never> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
	expand?: T
}
// ...
export type EntriesRecord<Tmeta = unknown> = {
	category: RecordIdString
	content: HTMLString
	created?: IsoDateString
	id: string
	media?: string[]
	meta?: null | Tmeta
	slug: string
	title: string
	updated?: IsoDateString
}
// ...
export type EntriesResponse<Tmeta = unknown, Texpand = unknown> = Required<EntriesRecord<Tmeta>> & BaseSystemFields<Texpand>